Two minor fixes to avoid compiler warnings.
git-svn-id: svn://svn.code.sf.net/p/rdesktop/code/trunk/rdesktop@1477 423420c4-83ab-492f-b58f-81f9feb106b5

+++ pstcache.c
@@ -71,7 +71,8 @@ pstcache_load_bitmap(uint8 cache_id, uint16 cache_idx)
 	rd_read_file(fd, celldata, cellhdr.length);
 
 	bitmap = ui_create_bitmap(cellhdr.width, cellhdr.height, celldata);
-	DEBUG(("Load bitmap from disk: id=%d, idx=%d, bmp=0x%x)\n", cache_id, cache_idx, bitmap));
+	DEBUG(("Load bitmap from disk: id=%d, idx=%d, bmp=0x%x)\n", cache_id, cache_idx,
+	       (unsigned int) bitmap));
 	cache_put_bitmap(cache_id, cache_idx, bitmap);
 
 	xfree(celldata);
